utils/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["shin5ok <shin5ok@55mp.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-firestore = "^2.6.1"
click = "^8.1.3"
Faker = "^14.1.0"
google-cloud-datastore = "^2.20.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
